ponderada_03/back/img_filter/requirements.txt
flask[async]
Pillow